requirements.txt
Django
djangorestframework
requests
django-environ
django-cors-headers 
drf-yasg
psycopg2
google-generativeai